move benchmark commands to archiver.benchmarks
